+++ sqllogictest-engines/src/external.rs
@@ -1,6 +1,6 @@
 use std::io;
 use std::marker::PhantomData;
-use std::process::Stdio;
+use std::process::{ExitStatus, Stdio};
 use std::time::Duration;
 
 use async_trait::async_trait;
@@ -119,6 +119,10 @@ impl AsyncDB for ExternalDriver {
 
     async fn sleep(dur: Duration) {
         tokio::time::sleep(dur).await
+    }
+
+    async fn run_command(command: std::process::Command) -> std::io::Result<ExitStatus> {
+        Command::from(command).status().await
     }
 }