+++ sqllogictest-engines/src/postgres/extended.rs
@@ -1,4 +1,5 @@
 use std::fmt::Write;
+use std::process::{Command, ExitStatus};
 use std::time::Duration;
 
 use async_trait::async_trait;
@@ -316,5 +317,9 @@ impl sqllogictest::AsyncDB for Postgres<Extended> {
 
     async fn sleep(dur: Duration) {
         tokio::time::sleep(dur).await
+    }
+
+    async fn run_command(command: Command) -> std::io::Result<ExitStatus> {
+        tokio::process::Command::from(command).status().await
     }
 }